[TASK] basis comparison of lattices: useful to see if exported matches original
